samples/requirements.txt
python-dotenv
httpx
pydantic
asyncio